+++ ddosdiversion-20230701/ChangeLog.md
@@ -1,3 +1,8 @@
+2024-04-28 Version: 1.0.1
+- Update API ListInstance: update response param.
+- Update API QueryNetList: update response param.
+
+
 2023-12-13 Version: 1.0.0
 - Generated python 2023-07-01 for DdosDiversion.